+++ components/reformeur/impact/bar-chart.jsx
